+++ VirtualShop.Domai/Entities/Products.cs
@@ -5,6 +5,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using VirtualShop.Domain.Commands;
 using VirtualShop.Domain.Validations;
 
 namespace VirtualShop.Domain.Entities;
@@ -13,10 +14,10 @@ public class Products
 {
     public int Id { get; set; }
     public string Description { get; set; }
-    public double UnitPrice { get; set; }
+    public decimal UnitPrice { get; set; }
     public double StockQuantity { get; set; }
 
-    public Products(int id, string description, double unitPrice, double stockQuantity)
+    public Products(int id, string description, decimal unitPrice, double stockQuantity)
     {
         Id = id;
         Description = description;
@@ -31,5 +32,5 @@ public class Products
     public Products(int id)
     {
         Id = id;
-    }
+    } 
 }
